aurora_backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
scikit-learn>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.1
torch>=2.1.0
# SpoonOS core developer framework (graph+agents+tools)
spoon-core @ git+https://github.com/XSpoonAi/spoon-core.git@main
